Add support for simple and advanced stats inside of ndt
